pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "casual-mcp"
version = "0.8.0"
description = "Multi-server MCP client for LLM tool orchestration"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Alex Stansfield" }]
dependencies = [
    "casual-llm[openai]>=0.6.0",
    "dateparser>=1.2.1",
    "fastapi>=0.115.12",
    "fastmcp>=2.14.5,<3.0.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.0",
    "questionary>=2.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "typer>=0.19.2",
    "uvicorn>=0.34.2",
]

[project.urls]
"Homepage" = "https://github.com/casualgenius/casual-mcp"
"Repository" = "https://github.com/casualgenius/casual-mcp"
"Issue Tracker" = "https://github.com/casualgenius/casual-mcp/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
casual-mcp = "casual_mcp.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "black>=25.1.0",
    "coverage>=7.8.0",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.11.10",
]

[tool.bumpversion]
current_version = "0.8.0"
commit = true
tag = true

[tool.bumpversion.file."pyproject.toml"]
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'
